requirements.txt
camelot_py==0.9.0
docx2pdf==0.1.8
fitz==0.0.1.dev2
ollama==0.2.1
langchain==0.3.1
langchain_community==0.3.1
langchain_text_splitters==0.3.0
matplotlib==3.8.2
numpy==2.1.1
pandas==2.2.3
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.10
python-dotenv==1.0.1
pywin32==306
streamlit==1.36.0
